pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-triage"
version = "0.1.0"
description = "layered e-mail classification: cache, embedding, heuristics and LLM."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "typer",
    "python-dotenv",
    "sentence-transformers",
    "numpy",
    "pyyaml",
    "ollama",
]

[project.scripts]
triage = "triage.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["triage*"]
